requirements-dev.txt
-r requirements.txt
flake8>=2.2.4
pytest-flask==0.7.5
factory-boy==2.4.1
